Fix -Wstar-is-type warnings
As of GHC 9.0.1, `-Wall` enables `-Wstar-is-type`, which warns if `*` is used
as a kind instead of `Type` (from `Data.Kind`). This patch fixes two
`-Wstar-is-type` warnings in `Language.Go.AST`.

Since `Data.Kind` has only been around since GHC 8.0, I bumped the lower
version bounds on `base` to `>= 4.9` to ensure that `Data.Kind` is always
available.